+++ src/dom/useTypedParams.ts
@@ -2,7 +2,7 @@ import { Route, OutParams } from "../common/index.js";
 import { useParams } from "react-router-dom";
 import { useMemo } from "react";
 
-export function useTypedParams<TPath extends string, TPathTypes, TSearchTypes, THash extends string[], TStateParams>(
+export function useTypedParams<TPath extends string, TPathTypes, TSearchTypes, THash extends string, TStateParams>(
     route: Route<TPath, TPathTypes, TSearchTypes, THash, TStateParams>
 ): OutParams<TPath, TPathTypes> {
     const params = useParams();