+++ src/dom/useTypedState.ts
@@ -2,7 +2,7 @@ import { Route, OutStateParams } from "../common/index.js";
 import { useLocation } from "react-router-dom";
 import { useMemo } from "react";
 
-export function useTypedState<TPath extends string, TPathTypes, TSearchTypes, THash extends string[], TStateParams>(
+export function useTypedState<TPath extends string, TPathTypes, TSearchTypes, THash extends string, TStateParams>(
     route: Route<TPath, TPathTypes, TSearchTypes, THash, TStateParams>
 ): OutStateParams<TStateParams> {
     const location = useLocation();