+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+-   [UNFINISHED] Hash values can now be specified without the `hashValues()` helper.
+-   **Breaking**: In corresponding generics, the `THash` constraint is changed from `string[]` to `string`.
+
 ## [1.2.2] - 2024-04-21
 
 ### Fixed
@@ -118,6 +125,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Hook dependencies are now properly listed, which is checked by ESLint. This fixes `useTypedSearchParams` for dynamic routes.
 -   Prevent access to internal `useUpdatingRef` helper.
 
+[unreleased]: https://github.com/fenok/react-router-typesafe-routes/tree/dev
 [1.2.2]: https://github.com/fenok/react-router-typesafe-routes/tree/v1.2.2
 [1.2.1]: https://github.com/fenok/react-router-typesafe-routes/tree/v1.2.1
 [1.2.0]: https://github.com/fenok/react-router-typesafe-routes/tree/v1.2.0